refactor(prosemirror): rename package oak-addon-richtext-field-prosemirror
