+++ .storybook/index.styl
@@ -1,7 +1,7 @@
 @import "../packages/oak/lib/index"
 @import "../packages/oak-addon-basic-components/lib/index"
 @import "../packages/oak-addon-richtext-field/lib/index"
-@import "../packages/oak-prosemirror-editor/lib/index"
+@import "../packagesoak-addon-richtext-field-prosemirror/lib/index"
 
 body
   padding: 20px 0
