+++ packages/oak-addon-richtext-field-prosemirror/lib/core/Editor/Toolbar/ColorButton.js
