+++ packages/oak-addon-richtext-field-prosemirror/lib/core/Editor/index.js
