+++ packages/oak-addon-richtext-field-prosemirror/lib/core/Editor/schema.js
